pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "osm-raster-topology"
version = "0.1.0"
description = "Design scaffold for topology-aware OSM to raster conversion."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Codex" }]
dependencies = [
  "numpy>=1.26",
  "pillow>=10.0",
]

[project.optional-dependencies]
runtime = [
  "networkx>=3.2",
  "shapely>=2.0",
  "rasterio>=1.3",
  "pyosmium>=4.0",
]
validation = [
  "scikit-image>=0.22",
]

[project.scripts]
osm-topology = "osm_raster_topology.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
